requirements.txt
streamlit
pandas
numpy
scikit-learn
openai
python-dotenv
plotly
nest_asyncio
polars
weasyprint
chardet
openpyxl
flask